refacto(cli): Create dedicated file to create the config for commands.
